Resolve a duplicate static symbol during Unity Build
Removed a duplicate function, and replaced it with direct calculation
since it was only being used once in one of the files anyway. The same
for `invLog2`, as the compiler should be taking care of it already.

Task-number: QTBUG-109394
Change-Id: I9095bca32ea127f770fa2cb0523b05110bd8ef21
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 95e2cae51e6c92315fbf33c5d3e8a1826cca8907)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>