Doc: Fix links and snippets
Add modules in which linked-to types live to the qdoc dependencies, fix
links, and adapt snippets to changes in the examples (and add tags to
quoted files as needed).

This makes the documentation generate without warnings, so set the
warninglimit to 0 as well to block future regressions.

Change-Id: I43fdfce1087c2a35212f91b41bba1ff9481e609d
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
Reviewed-by: <safiyyah.moosa@qt.io>
(cherry picked from commit fea04eea8b142aa5ac89b1024bbfc5298685fb0c)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>