MapView: add bearing / tilt wheel functionality; also with Magic Mouse
In Qt 5, Map had bearing / tilt mouse wheel functionality with keyboard
modifiers; that feature is now restored. We allow the wheel to zoom, to
change the bearing if the Shift key is held, or to change the tilt if
the Control key is held.

This is redundant when you're using a touchpad that can provide the
pinch-zoom feature; but we need it with the Magic Mouse on macOS,
because the cocoa plugin can't distinguish it from a trackpad, even
though it can't provide the pinch gesture. So we enable these features
for both mouse and touchpad devices on macOS.

Also allow mouse and touchpad with the minimal_map example.

There's a similar issue on Wayland, but let's try to fix the Wayland
plugin to distinguish these devices properly.

Task-number: QTBUG-87646
Task-number: QTBUG-112394
Change-Id: I9e516a52e941b0b05cd0114afafde2767d80f4ff
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 13d2a4e31dbef2b48ebc44978368539ba96fded8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>