Allow zoom/pan/tilt with mouse on Wayland
This is redundant when you're using a touchpad that can provide the
pinch-zoom feature; but so far we are not distinguishing mice and
touchpads on Wayland. So in case the device that claims to be a touchpad
is actually just a mouse, the user needs some way to zoom, pan and tilt.
This can be reverted when QTBUG-112432 is fixed.

Pick-to: 6.5
Task-number: QTBUG-112394
Task-number: QTBUG-112432
Change-Id: Ie339e3c1d7d0b1d5b6383d8e7f8fd1cdcb462db6
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>