CoreLocation plugin: introduce RequestAlwaysPermission parameter
Previously we were using NSLocationAlwaysAndWhenInUseUsageDescription
and NSLocationWhenInUseUsageDescription to distinguish the level of
permission that we want to request.
However, Apple now always requires *both* entries to present in the
Info.plist file.

[ChangeLog][Important Behavior Changes][Apple] Introduce an explicit
RequestAlwaysPermission parameter to the corelocation plugin.
When specified and set to true, this parameters enables the "Always"
location request. When not specified or set to false, only the
"When In Use" location permission is requested. By default this
parameter is not specified.

This patch is a manual cherry-pick from Qt 6 qtpositioning repo
(commit 697f2dbb559966a0f7aeccdb5a905da39ebcfab6).
In Qt 5 this change also requires to update the
QGeoPositionInfoSourceFactoryCL, so that it now inherits from
QGeoPositionInfoSourceFactoryV2. That is required to pass parameters
to the plugin.

Fixes: QTBUG-109359
Change-Id: Ib35f578a6459e84d2ff29c79416ae5ee9f104cc8
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 697f2dbb559966a0f7aeccdb5a905da39ebcfab6)