Tag some examples with example categories
Most examples are obviously in the "Positioning & Location" category.
The mapviewer is a good fit for the "Mobile" category as the layout, UI,
and functional integration with positioning makes it particularly useful
on mobile platforms.

Some examples remain uncategorised on purpose, as they show a very
specific subset of the framework.

Pick-to: 6.5 6.6
Change-Id: I077b706856f8a480d4d04517e65ddff47a0c56aa
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Kimmo Leppälä <kimmo.leppala@qt.io>
Reviewed-by: Safiyyah Moosa <safiyyah.moosa@qt.io>