Add GeoJson QML API and refactor GeoJson Viewer example
The GeoJson Viewer example used private C++ functions to open and save
GeoJson files. QtLocation is not intended to have a public C++ API and
it is therefore required to create a QML API to make these functions
available to our users.

This patch moves some code of the example, responsible for exposing the
C++ GeoJson functions to QML, into the QtLocation source code. The new
type GeoJsonData can be used to load and save GeoJson files into a
model, which can then be rendered using delegates.

Change-Id: Icce78f48d77f303faa96ac065ad96025eded83e7
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>