Prevent QGeoManeuver from being exported to QML twice
We need to export the type as a value type, and also export its enums
in a namespace. This resulted in duplicated entries in the qmltypes
file, which confuses Qt Design Studio.

Fix it by introducing a derived class, and using it for exporting the
enums.

Fixes: QTBUG-115359
Change-Id: I499134218dccdede942cfad5e205ce598ef28b66
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 3e9facacf5a6dabb1ac2990b65bd80b09974284e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>