WeatherInfo example: fix a potential crash
The positionError() slot was unconditionally deleting the src object.
However, in practice, when starting the example with Location services
disabled, we might receive multiple errors, so the slot will be
triggered multiple times.
As a result, the second call of the slot will try to dereference a
nullptr.
Fix it by explicitly checking src for nullptr.

Change-Id: I4bbe12154ece80529b4dd4c5d8eaab079ad5c18a
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>