places_map example: add support for LocationPermission
The example didn't make use of the new permissions API, and, as a
result, it was throwing errors when trying to use PositionSource.

Rework the example to use the LocationPermission API.
The idea is that the premission is checked during the application
startup. If the permission is not granted, a new window is shown. This
window contans some explanations on why the user should enable the
location permission and a "button" to do that.
If the permission is granted, the Loaded is used to load a normal
application page. Otherwise an error message is displayed.

Task-number: QTBUG-121412
Pick-to: 6.7 6.6
Change-Id: I992e485b71b193e4a106cf42aed8c17f57b9fc46
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>