MapViewer Example: Support LocationPermission
If the location permission is not granted, a message will be shown to
inform the user that this app needs location permission which can be
requested by the user or maybe not.
If the user triggers "Follow Me" action, then it checks if the
permission has already Granted, then toggles the follow-me property.
Otherwise, it shows the permission request dialog.

Task-number: QTBUG-121412
Pick-to: 6.6 6.7
Change-Id: Ia6cf7a11e6ce9b45c66f841ea027ed78fc2beda7
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>