Fix GeoJson Viewer example Android deployment
As of 6.7.0, androiddeployqt doesn't deploy plugins to the Android .apk
if they are not explicitly linked to the project. The missins plugins
cause the example to not load the maps and to be rather useless.

Let's fix this by adding Location / Qt::Location to CMakeLists.txt.

Fixes: QTBUG-123112
Pick-to: 6.7.0
Change-Id: I67ddc67f58fd9eb2e7225a336d559d1dd2e91c36
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit a62a17e3718f8cffd4a85fef4a1472765040d042)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>