tst_map_keepgrab: remove tryCompare(flickEndedSpy, "count", 0)
The line that tested if the outer flickable didn't stop flicking
when dragging on the inner MapView, was wrong. The outer
flickable will stop flicking in that case (the content item stops
moving), it's just that the "flicking" property is not
updated. And that is a bug, not a feature.

In order to fix that bug in QtDeclarative, we first need to
remove this faulty test line, since it will otherwise start
to fail once we correct the behavior in the other
repository.

Task-number: QTBUG-130139
Change-Id: I2f8a9e8dc0443dcef7b7d7f6a4584add3dd29884
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit ff820634973f82b1aa0a66cb82f9ec6c96437faa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d18337c5d754bc0c043c036339ef68a379f6db9a)