tst_map_keepgrab: remove tryCompare(flickEndedSpy, "count", 0)
The line that tested if the outer flickable didn't stop flicking
when dragging on the inner MapView, was wrong. The outer
flickable will stop flicking in that case (the content item stops
moving), it's just that the "flicking" property is not
updated. And that is a bug, not a feature.

In order to fix that bug in QtDeclarative, we first need to
remove this faulty test line, since it will otherwise start
to fail once we correct the behavior in the other
repository.

Task-number: QTBUG-130139
Pick-to: 6.8 6.5
Change-Id: I2f8a9e8dc0443dcef7b7d7f6a4584add3dd29884
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>