Add contains() function to QDeclarativeGeoMapQuickItem
The default QQuickItem::contains() does not work for this class because
this class uses some non-default transformations in case of
zoomLevel != 0.

This patch tries to solve the issue by transforming the incoming point
into the correct coordinate system before calling
QQuickItem::contains(). The mathematical principle is taken from the
implementation of QDeclarativeGeoMap::fitViewportToMapItemsRefine.

Be aware that the property zoomLevel is untested and not fully
implemented, see the many TODOs in the respective code.

Fixes: QTBUG-134438
Change-Id: Ia5d85d4e3a64be4975ec08a91a55ab00c825f171
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Lena Biliaieva <lena.biliaieva@qt.io>