Use the transformed path to check for contains
QDeclarativePolylineMapItem contains two paths, m_geometry.srcPath_
and m_shapePath->path(). The former is untransformed, the latter is
transformed. To calculate if a point is contained by the path the latter
has to be used.

Further, this patch removes an offset to the point that was wrong.

Fixes: QTBUG-134097
Pick-to: 6.8
Change-Id: I7dbc0ff2ad9bd9b37570c838825260dbf9aa9ffb
Reviewed-by: Mate Barany <mate.barany@qt.io>
(cherry picked from commit 0c4874978257ed7da7c06992d907a440be8ab498)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>